routes refactoring
